+++ src/Terrabuild.Configuration/AST/Project.fs
@@ -29,6 +29,7 @@ type TargetBlock =
       Cache: Expr option
       Batch: Expr option
       Phase: Expr option
+      Lock: Expr option
       EnvironmentSensitive: Expr option
       Steps: Step list }