+++ src/Terrabuild.Configuration/Transpiler/Project.fs
@@ -75,7 +75,7 @@ let toProject (block: Block) =
 
 let toTarget (block: Block) =
     block
-    |> checkAllowedAttributes ["outputs"; "depends_on"; "build"; "artifacts"; "batch"; "phase"; "environment_sensitive"]
+    |> checkAllowedAttributes ["outputs"; "depends_on"; "build"; "artifacts"; "batch"; "phase"; "lock"; "environment_sensitive"]
     |> ignore
 
     let outputs = block |> tryFindAttribute "outputs"
@@ -91,6 +91,7 @@ let toTarget (block: Block) =
     let cache = block |> tryFindAttribute "artifacts"
     let batch = block |> tryFindAttribute "batch"
     let phase = block |> tryFindAttribute "phase"
+    let targetLock = block |> tryFindAttribute "lock"
     let environmentSensitive = block |> tryFindAttribute "environment_sensitive"
     let steps =
         block.Blocks
@@ -119,6 +120,7 @@ let toTarget (block: Block) =
       TargetBlock.Cache = cache
       TargetBlock.Batch = batch
       TargetBlock.Phase = phase
+      TargetBlock.Lock = targetLock
       TargetBlock.EnvironmentSensitive = environmentSensitive
       TargetBlock.Steps = steps }