+++ src/Terrabuild/Core/GraphPipeline/Batch.fs
@@ -454,6 +454,7 @@ let private createBatchNodes (options: ConfigOptions.Options) (configuration: Co
                   GraphDef.Node.ProjectDir = "."
                   GraphDef.Node.Target = headNode.Target
                   GraphDef.Node.Phase = batch.Phase
+                  GraphDef.Node.Locks = batch.Nodes |> Seq.collect _.Locks |> Set.ofSeq
                   GraphDef.Node.Operations = ops
                   GraphDef.Node.EvaluationInputs =
                     batch.Nodes